Prevent out of memory errors
The program frequently crashed the server when reading in the bigger
id files. This commit frequently "flushes" the tweets to disk, removing
the need to keep them in memory.

Because there can now exist partially written files, the ".tmp"
extension is added while the rehydration of a file is in progress, and
removed once the tweets are all finished.

Other minor changes:

(1) Out files now use the '.csv' extension.

(2) The out file naming scheme has been changed to be only the date,
    instead of starting with 'bth_ids' like the in files.

(3) A status update is printed every 100 tweets processed.

(4) The csv writer always uses '\n', for consistency across platforms.
    This change also saves a byte per tweet by omitting the '\r'.

(5) Type annotations are added to all top-level methods.