Print status when determining unhydrated tweets
Finding which tweets have been hydrated is a length process if the in
file is large, which leaves the user with no idea of if the program is
running. This is further compounded by the fact that the check happens
at the beginning of the program execution.

Simply printing out the program's status is a great aid to user
friendliness.